refactor: unify FSM core implementation
